move compression utilities into own package
